+++ evalml/tests/pipeline_tests/test_component_graph.py
@@ -2767,6 +2767,6 @@ def test_get_component_input_logical_types():
     no_estimator.fit(X, y)
     assert no_estimator.last_component_input_logical_types == {
         "cat": Categorical(),
-        "numeric": Integer(),
+        "numeric": Double(),
         "email": EmailAddress(),
     }
